+++ src/routes/about/members/+page.server.ts
@@ -1,18 +1,6 @@
 import type { Person } from 'schema-dts';
-import createClient from '@sanity/client';
-import imageUrlBuilder from '@sanity/image-url';
-import { client } from '../../../gql-client';
+import { graphQlClient, imageUrlBuilder } from '../../../clients';
 import { graphql } from '../../../gql';
-
-export const sanityClient = createClient({
-	projectId: '95rc74zt',
-	dataset: 'production',
-	useCdn: false,
-	apiVersion: '2022-01-12',
-	token: process.env.SANITY_TOKEN
-});
-
-const builder = imageUrlBuilder(sanityClient);
 
 export const prerender = true;
 
@@ -81,30 +69,30 @@ const getMembersDocument = graphql(`
 	}
 `);
 
-export interface LoadOutput {
+interface LoadOutput {
 	members: Omit<Person, '@type'>[];
 }
 
-export function notEmpty<T>(value: T | null | undefined): value is T {
+function notEmpty<T>(value: T | null | undefined): value is T {
 	return value !== null && value !== undefined;
 }
 
 export const load = async (): Promise<LoadOutput> => {
-	const { members } = await client.request(getMembersDocument);
+	const { members } = await graphQlClient.request(getMembersDocument);
 
 	const output = members
 		.map(({ member }) => {
 			const person = member?.person;
 
 			const image = person?.image
-				? builder.image(person.image).auto('format').width(200).height(200).toString()
+				? imageUrlBuilder.image(person.image).auto('format').width(200).height(200).toString()
 				: undefined;
 
 			return {
 				givenName: person?.givenName,
 				familyName: person?.familyName,
 				email: person?.email,
-				phone: person?.phone,
+				telephone: person?.phone,
 				address: person?.address,
 				image,
 				worksFor: person?.worksFor
