+++ src/routes/about/members/member.svelte
@@ -46,11 +46,20 @@
 			<h2>{name}</h2>
 			<div class="roles">
 				{#each worksFor as role}
-					<p>{role.name} ved {role.worksFor.name}</p>
+					<p>
+						{role.name} ved
+						{#if role.worksFor?.sameAs && Array.isArray(role.worksFor.sameAs) && role.worksFor.sameAs.length > 0 && !role.worksFor.sameAs[0].includes('klyngeorg.no')}
+							<a href={role.worksFor.sameAs[0]} target="_blank" rel="noreferrer">
+								{role.worksFor.name}
+							</a>
+						{:else}
+							{role.worksFor.name}
+						{/if}
+					</p>
 				{/each}
 			</div>
 		</header>
-		<div>
+		<div class="contact-details">
 			{#if data.telephone}
 				<p><a href={`tel:${data.telephone}`}>{data.telephone}</a></p>
 			{/if}
@@ -75,8 +84,13 @@
 		gap: var(--space-xs);
 	}
 
-	.roles {
+	.roles,
+	.contact-details {
 		font-size: var(--font-size-sm);
+	}
+
+	.roles a {
+		text-decoration: none;
 	}
 
 	article p {
